Code cleanup (module interceptor) + new config to create pcap file
