+++ src/config.rs
@@ -1,8 +1,10 @@
 use clap::{crate_name, crate_version, App, Arg};
 
+#[derive(Clone)]
 pub struct Config {
     listener_url: String,
     routing_url: Option<String>,
+    pcap_filename: Option<String>,
 }
 
 impl Config {
@@ -12,6 +14,10 @@ impl Config {
 
     pub fn routing_url(&self) -> Option<String> {
         self.routing_url.clone()
+    }
+
+    pub fn pcap_filename(&self) -> Option<String> {
+        self.pcap_filename.clone()
     }
 
     pub fn init() -> Self {
@@ -40,6 +46,16 @@ impl Config {
                     .long_help("An address on which the server will route all packets. Format: <scheme>://<ip>:<port>. Scheme supported : tcp and tls. If it is not specified, the JET protocol will be used.")
                     .takes_value(true)
                     .empty_values(false),
+            )
+            .arg(
+                Arg::with_name("pcap-filename")
+                    .short("p")
+                    .long("pcap")
+                    .value_name("PCAP_FILENAME")
+                    .help("Path of the file where the pcap file will be saved. If not set, no pcap file will be created.")
+                    .long_help("Path of the file where the pcap file will be saved. If not set, no pcap file will be created.")
+                    .takes_value(true)
+                    .empty_values(false),
             );
 
         let matches = cli_app.get_matches();
@@ -48,9 +64,12 @@ impl Config {
 
         let routing_url = matches.value_of("routing-url").map(|url| url.to_string());
 
+        let pcap_filename = matches.value_of("pcap-filename").map(|pcap_filename| pcap_filename.to_string());
+
         Config {
             listener_url,
             routing_url,
+            pcap_filename,
         }
     }
 }
