+++ src/routing_client.rs
@@ -6,17 +6,20 @@ use url::Url;
 
 use crate::transport::tcp::TcpTransport;
 use crate::transport::Transport;
-use crate::build_proxy;
+use crate::config::Config;
+use crate::Proxy;
 
 pub struct Client {
     routing_url: Url,
+    config: Config,
     _executor_handle: TaskExecutor,
 }
 
 impl Client {
-    pub fn new(routing_url: Url, executor_handle: TaskExecutor) -> Self {
+    pub fn new(routing_url: Url, config: Config, executor_handle: TaskExecutor) -> Self {
         Client {
             routing_url,
+            config,
             _executor_handle: executor_handle,
         }
     }
@@ -28,7 +31,7 @@ impl Client {
         let server_conn = TcpTransport::connect(&self.routing_url);
 
         Box::new(server_conn.and_then(move |server_transport| {
-            build_proxy(server_transport, client_transport)
+            Proxy::new(self.config.clone()).build(server_transport, client_transport)
         }))
     }
 }
